iniciando re setup
