+++ package.json
@@ -1,31 +1,30 @@
 {
   "name": "ic819-backend",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "express": "^4.18.1",
-    "express-session": "^1.17.3",
-    "express-session-mariadb-store": "^0.1.4",
-    "jest": "^28.1.3",
-    "mariadb": "^3.0.1",
-    "passport-jwt": "^4.0.0",
-    "sequelize": "^5.21.3"
-  },
+  "version": "1.0.0",
+  "description": "",
+  "main": "server.js",
   "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
     "dev": "nodemon ./src/server.js"
   },
-  "eslintConfig": {
-    "extends": [
-    ]
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/taernsietr/IC819-backend.git"
+  },
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/taernsietr/IC819-backend/issues"
+  },
+  "homepage": "https://github.com/taernsietr/IC819-backend#readme",
+  "dependencies": {
+    "dotenv": "^16.0.2",
+    "express": "^4.18.1"
   },
   "devDependencies": {
-    "@types/express": "^4.17.13",
-    "@types/node": "^18.6.4",
-    "@typescript-eslint/eslint-plugin": "^5.30.7",
-    "@typescript-eslint/parser": "^5.30.7",
-    "eslint": "^8.20.0",
-    "nodemon": "^2.0.19",
-    "ts-node": "^10.9.1",
-    "typescript": "^4.7.4"
+    "eslint": "^8.23.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-plugin-import": "^2.26.0",
+    "nodemon": "^2.0.19"
   }
 }
