+++ src/controllers/clientController/clientController.js
@@ -1,0 +1,13 @@
+import { responseCodes } from "../../resources";
+
+async function createClient(req, res) {
+	try {
+		const { data } = req.body;
+
+		// validar os dados
+
+		
+	} catch (error) {
+
+	}
+}
